Refactor: new obs_controller
Loading distribution information logic from OBS is moved to a separate
controller along with methods that depend on this kind of information.

If the connection to OBS is down, searching is disabled and the user
is notified.

.rubocop_todo.yml was updated because offending code was moved
