+++ app/controllers/distributions_controller.rb
@@ -1,4 +1,4 @@
-class DistributionsController < ApplicationController
+class DistributionsController < OBSController
   before_action :set_releases_parameters, only: %i[index testing leap]
 
   # GET /distributions
